requirements.txt
langchain
langchain-community
langchain-ollama
langchain-chroma
chromadb
fastapi
uvicorn
python-multipart
python-dotenv
langchain-groq
langchain-huggingface
sentence-transformers
docx2txt
pandas
openpyxl
